Pitest will not fail on empty modules
Some, currently one one module, does not have any code. The default
behavior of Pitest is to fail, if no mutations have been found.

Pitest was reconfigured not to fail if there are no mutations.
This might only happen in this project if there is no code
at all in a module.